+++ lib/loopback.js
@@ -120,13 +120,16 @@ if (loopback.isServer) {
 
 if (loopback.isServer) {
   fs
-    .readdirSync(path.join(__dirname, 'middleware'))
+    .readdirSync(path.join(__dirname, '..', 'server', 'middleware'))
     .filter(function(file) {
       return file.match(/\.js$/);
     })
     .forEach(function(m) {
-      loopback[m.replace(/\.js$/, '')] = require('./middleware/' + m);
+      loopback[m.replace(/\.js$/, '')] = require('../server/middleware/' + m);
     });
+
+  loopback.urlNotFound = loopback['url-not-found'];
+  delete loopback['url-not-found'];
 }
 
 /*
