+++ server/middleware/static.js
